Object for plotting commonality analyses

Object for plotting commonality analyses

ggcommonality(
  formula,
  data,
  add_ci = TRUE,
  stack = FALSE,
  stack_by = "common",
  n_replications = 100,
  sample_column = NULL,
  resample_type = "wild",
  wild_type = "gaussian",
  include_total = FALSE,
  seed = NULL
)

Arguments

formula

Formula in form of y ~ x1 + x2

data

Data.frame object containing data to be visualized

add_ci

Logical. Add bootstrap-estimated confidence interval?

stack

Logical. Stack commonality effects?

stack_by

Character. Either "common" to stack unique vs. common effects or "partition" to stack unique and common effects for each IV.

n_replications

Numeric. Number of replications for bootstrap simulation.

sample_column

Character. Name of column to perform stratified sampling with, or leave as NULL

resample_type

Character. Method for boostrap resampling. Either "random", "fixed", or "wild". See README for details.

wild_type

Character. If resample_type == "wild", either "Gaussian" to multiply resampled residuals by random constants from the normal distribution, or sign to randomly multiply half of the residuals by +1 and half by -1. This provides a solution to "fixed" in the presence of model heteroscedasticity. See README for details

include_total

Logical. Include bar representing total variance explained across all unique and common effects?

seed

Numeric. Number to set R's randomization seed to (for reproducibility).